+++ fs/src/lib.rs
@@ -1,0 +1,36 @@
+//! Cuprate directories and filenames.
+//!
+//! # Environment variables on Linux
+//! Note that this module's functions uses [`dirs`],
+//! which adheres to the XDG standard on Linux.
+//!
+//! This means that the values returned by these statics
+//! may change at runtime depending on environment variables,
+//! for example:
+//!
+//! By default the config directory is `~/.config`, however
+//! if `$XDG_CONFIG_HOME` is set to something, that will be
+//! used instead.
+//!
+//! ```rust
+//! # use cuprate_helper::fs::*;
+//! # if cfg!(target_os = "linux") {
+//! std::env::set_var("XDG_CONFIG_HOME", "/custom/path");
+//! assert_eq!(
+//!     CUPRATE_CONFIG_DIR.to_string_lossy(),
+//!     "/custom/path/cuprate"
+//! );
+//! # }
+//! ```
+//!
+//! Reference:
+//! - <https://github.com/Cuprate/cuprate/issues/46>
+//! - <https://docs.rs/dirs>
+
+mod constants;
+mod network;
+mod paths;
+
+pub use constants::{CUPRATE_DIR, DEFAULT_CONFIG_FILE_NAME};
+pub use network::{address_book_path, arti_path, blockchain_path, logs_path, txpool_path};
+pub use paths::{CUPRATE_CACHE_DIR, CUPRATE_CONFIG_DIR, CUPRATE_DATA_DIR};
